src/main/java/tk/youngdk/datajpa/repository/MemberJpaRepository.java
package tk.youngdk.datajpa.repository;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import tk.youngdk.datajpa.domain.Member;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
@RequiredArgsConstructor
public class MemberJpaRepository {
    private final EntityManager em;

    public Long save(Member member) {
        em.persist(member);
        return member.getId();
    }

    public Member find(Long memberId) {
        Member findMember = em.find(Member.class, memberId);
        return findMember;
    }
}
