src/test/java/tk/youngdk/datajpa/repository/MemberJpaRepositoryTest.java
package tk.youngdk.datajpa.repository;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Commit;
import org.springframework.test.annotation.Rollback;
import org.springframework.transaction.annotation.Transactional;
import tk.youngdk.datajpa.domain.Member;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional(readOnly = true)
@Rollback(false)
class MemberJpaRepositoryTest {

    @Autowired
    MemberJpaRepository memberJpaRepository;

    @Test
    @DisplayName("저장 테스트")
    @Transactional
    public void save() throws Exception {
        //given
        Member member = new Member("memberA");

        //when
        Long savedMemberId = memberJpaRepository.save(member);
        Member findMember = memberJpaRepository.find(savedMemberId);

        //then
        assertThat(savedMemberId).isEqualTo(member.getId());
        assertThat(findMember.getUserName()).isEqualTo(member.getUserName());
        assertThat(findMember).isEqualTo(member);

    }

    @Test
    void find() {
    }
}